requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
audioop-lts==0.2.1
bitarray==3.1.1
cachetools==4.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
cryptography==44.0.2
cytoolz==1.0.1
dateparser==1.2.1
decorator==5.2.1
discord.py==2.5.2
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.3
pandas==2.2.3
parsimonious==0.10.0
pillow==11.1.0
propcache==0.3.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2025.1
pyunormalize==16.0.0
pywin32==309
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
toolz==1.0.0
tornado==6.4.2
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
